Fix undesired reordering of Drivers:
The FilterForCompatible function wasnt maintaining the
order of drivers that it started with. It was doing
this because goroutines would append their results as soon
as they finished without regard to order. Adding a state
map and surrounding it with mutex solves this concern and
leaves the drivers in the order they came in with.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>